Add enhancement of GUI layout (#6)
* Add mode play label and game status notification
