+++ src/players.c
@@ -1,4 +1,5 @@
 #include "main.h"
+#include "init.h"
 #include "players.h"
 #include "utils.h"
