+++ src/utils.c
@@ -1,4 +1,5 @@
 #include "main.h"
+#include "init.h"
 #include "players.h"
 #include "utils.h"
 
@@ -6,10 +7,7 @@ void checkEmptyPos(int boardEmptyPos[9][2]) {
     countEmptyPos = 0;
 
     for (int row = 0; row < 3; ++row) {
-        for (int col = 0; col < 3; ++col) {
-            boardEmptyPos[countEmptyPos][0] = 0;
-            boardEmptyPos[countEmptyPos][1] = 0;
-            
+        for (int col = 0; col < 3; ++col) {            
             if (board[row][col] == ' ') {
                 boardEmptyPos[countEmptyPos][0] = row;
                 boardEmptyPos[countEmptyPos][1] = col;
@@ -70,16 +68,23 @@ int checkWinnerOrDraw()
         if (*winner == *PLAYER1)
         {
             incCounterScore(p1Score);
+            gtk_label_set_label(noticeStatusLabel, "Player 1 Won");
         }
         else if (*winner == *PLAYER2)
         {
             incCounterScore(p2Score);
+            if (modeP1 == TRUE)
+                gtk_label_set_label(noticeStatusLabel, "Computer Won");
+            else
+                gtk_label_set_label(noticeStatusLabel, "Player 2 Won");
         }
         else
         {
             incCounterScore(tieScore);
+            gtk_label_set_label(noticeStatusLabel, "It's a tie");
         }
 
+        // gtk_widget_show(winDisplayModal);
         return 1;
     }
 
@@ -109,11 +114,19 @@ void setBoard(GtkGrid *grid)
     }
 }
 
-void setUp(GtkGrid *grid)
+void setUp()
 {
     strcpy(curPlayer, PLAYER1);
     *winner = ' ';
     blankSpaces = 9;
     reset = FALSE;
     setBoard(grid);
+}
+
+void resetBoard() {
+    setUp();
+    gtk_label_set_label(noticeStatusLabel, "Player 1's Turn");
+    gtk_label_set_text(p1Score, "0");
+    gtk_label_set_text(p2Score, "0");
+    gtk_label_set_text(tieScore, "0");
 }
